ci: only run ceph tests on linux

+++ ci/before_install.sh
@@ -14,9 +14,19 @@ if [[ "${TRAVIS_OS_NAME}" == "linux" ]]; then
 fi
 # osx has up-to-date cmake
 
-# start micro-osd with plugin
-pushd ${ZLOG_DIR}/docker/micro-osd
-docker build -t micro-osd .
-docker run --rm -d --net=host --volumes-from built-ceph-plugin \
-  -v /tmp/micro-osd:/micro-osd micro-osd
-popd
+# skip ceph tests on osx
+if [[ "${TRAVIS_OS_NAME}" == "linux" ]]; then
+  # build ceph plugin
+  pushd ${ZLOG_DIR}/docker/ceph-plugin
+  docker rm built-ceph-plugin || true
+  docker build -t ceph-plugin .
+  docker run -v /ceph_zlog_plugin --name built-ceph-plugin ceph-plugin
+  popd
+
+  # start micro-osd with plugin
+  pushd ${ZLOG_DIR}/docker/micro-osd
+  docker build -t micro-osd .
+  docker run --rm -d --net=host --volumes-from built-ceph-plugin \
+    -v /tmp/micro-osd:/micro-osd micro-osd
+  popd
+fi
